moved valueset expansion extensions to shims.base
